+++ Install/Chocolatey/tools/chocolateyInstall.ps1
@@ -1,8 +1,8 @@
 $packageName = 'markdownmonster'
 $fileType = 'exe'
-$url = 'https://github.com/RickStrahl/MarkdownMonsterReleases/raw/master/v1.23/MarkdownMonsterSetup-1.23.18.3.exe'
+$url = 'https://github.com/RickStrahl/MarkdownMonsterReleases/raw/master/v1.24/MarkdownMonsterSetup-1.24.8.exe'
 
 $silentArgs = '/VERYSILENT'
 $validExitCodes = @(0)
 
-Install-ChocolateyPackage "$packageName" "$fileType" "$silentArgs" "$url"  -validExitCodes  $validExitCodes  -checksum "BC6FDE4FE35A9206678E0DAA6E678F48D424670D44A7800D83D115F96E15191A" -checksumType "sha256"
+Install-ChocolateyPackage "$packageName" "$fileType" "$silentArgs" "$url"  -validExitCodes  $validExitCodes  -checksum "A8E30948D22448CC53E5F9F469CF26808049CE655DCD8D58E127A292B300ADB2" -checksumType "sha256"